web: don't import entire SourceViewPage in flow and user interface
